added getFloat functionality to cfg parsing
git-svn-id: file:///Users/braun/svn/vermont/branches/vermont/anom-detect@1557 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ cfg/RBSWormDetectorCfg.cpp
@@ -39,7 +39,7 @@ lambda_ratio(5)
 		} else if (e->matches("timecleanupinterval")) {
 			timeCleanupInterval = getInt("timecleanupinterval");
 		} else if (e->matches("lambdaratio")) {
-			lambda_ratio = (float) getInt("lambdaratio"); //no getfloat :(
+			lambda_ratio = getFloat("lambdaratio"); //no getfloat :(
 		} else if (e->matches("analyzerid")) {
 			analyzerId = e->getFirstText();
 		} else if (e->matches("idmeftemplate")) {
